Modified CBR deadline calculation, added CBRC option
Now calculates the next interval deadline based on the finishing time
of the previous interval. This is less accurate, and will not have
the 'catchup' behaviour in case of slowdowns. However, it is less
prone to issues with the system clock.

The previous behaviour is still available as CBRC, (CBR with catchup).